Added get user route and refactored functions into folder
