+++ functions/Login.ts
@@ -1,6 +1,6 @@
 import crypto from "crypto";
 import sqlite3 from "sqlite3";
-import { Name, Address } from "./types/interfaces";
+import { UserDbRow } from "../types/interfaces";
 
 const db = new sqlite3.Database("./db/users.db", (error) => {
   if (error) {
@@ -9,11 +9,9 @@ const db = new sqlite3.Database("./db/users.db", (error) => {
   console.log("Connected to the users database.");
 });
 
-export async function SignUpNewUser(
+export async function LoginUser(
   username: string,
   password: string,
-  name: Name,
-  address: Address
 ): Promise<number> {
   const hashedPassword = crypto
     .createHmac("sha256", "memes")
@@ -21,16 +19,21 @@ export async function SignUpNewUser(
     .digest("hex");
 
   return new Promise((resolve, reject) => {
-    db.run(
-      "INSERT INTO users (username,password,name,address) VALUES ($1,$2,$3,$4)",
-      [username, hashedPassword, JSON.stringify(name), JSON.stringify(address)],
-      (error) => {
+    db.get(
+      "SELECT * FROM users WHERE username=?",
+      [username],
+      (error, row: UserDbRow) => {
         if (error) {
           console.log(error);
           return reject(400);
         }
-        console.log("success");
-        return resolve(200);
+        if(row && row.password == hashedPassword){
+            console.log("success");
+            return resolve(200);
+        }else{
+            console.log("Error logging in");
+            return resolve(401);
+        }
       }
     );
   });
