+++ functions/NewUser.ts
@@ -1,6 +1,6 @@
 import crypto from "crypto";
 import sqlite3 from "sqlite3";
-import { UserDbRow } from "./types/interfaces";
+import { Name, Address } from "../types/interfaces";
 
 const db = new sqlite3.Database("./db/users.db", (error) => {
   if (error) {
@@ -9,9 +9,11 @@ const db = new sqlite3.Database("./db/users.db", (error) => {
   console.log("Connected to the users database.");
 });
 
-export async function LoginUser(
+export async function SignUpNewUser(
   username: string,
   password: string,
+  name: Name,
+  address: Address
 ): Promise<number> {
   const hashedPassword = crypto
     .createHmac("sha256", "memes")
@@ -19,21 +21,16 @@ export async function LoginUser(
     .digest("hex");
 
   return new Promise((resolve, reject) => {
-    db.get(
-      "SELECT * FROM users WHERE username=?",
-      [username],
-      (error, row: UserDbRow) => {
+    db.run(
+      "INSERT INTO users (username,password,name,address) VALUES ($1,$2,$3,$4)",
+      [username, hashedPassword, JSON.stringify(name), JSON.stringify(address)],
+      (error) => {
         if (error) {
           console.log(error);
           return reject(400);
         }
-        if(row && row.password == hashedPassword){
-            console.log("success");
-            return resolve(200);
-        }else{
-            console.log("Error logging in");
-            return resolve(401);
-        }
+        console.log("success");
+        return resolve(200);
       }
     );
   });
