+++ types/interfaces.ts
@@ -29,3 +29,8 @@ export interface LoginRequest {
     username: string,
     password: string
 }
+
+export interface UserInfo {
+    name: Name,
+    address: Address
+}
